Refactor KVCache to remove device_lengths and optimize memory operations for divergent sequence lengths

+++ src/core/generation.cu
@@ -94,7 +94,7 @@ std::vector<bpe::TokenId> generate_tokens(
     for (std::size_t step = 0; step < options.max_new_tokens; ++step) {
         const std::size_t begin = result.size() > options.max_context_tokens
             ? result.size() - options.max_context_tokens : 0;
-        const std::vector<bpe::TokenId> context(result.begin() + static_cast<size_t>(begin), result.end());
+        const std::vector<bpe::TokenId> context(result.begin() + static_cast<long long>(begin), result.end());
         Tensor logits({context.size(), model_options.vocabulary_size}, weights.token_embedding.dtype());
         TransformerModelWorkspace workspace(model_options, 1, context.size(), weights.token_embedding.dtype());
         bpe::TokenId* device_ids = nullptr;
@@ -106,7 +106,7 @@ std::vector<bpe::TokenId> generate_tokens(
             CUDA_CHECK(cudaStreamSynchronize(stream));
             std::vector<float> host_logits(logits.numel());
             logits.copy_to_host(host_logits);
-            std::vector<float> last_logits(host_logits.end() - model_options.vocabulary_size,
+            std::vector<float> last_logits(host_logits.end() - static_cast<long long>(model_options.vocabulary_size),
                                            host_logits.end());
             const bpe::TokenId next = sample(last_logits, options, rng);
             result.push_back(next);
